+++ packages/expo-video/ios/VideoPlayerItem.swift
@@ -28,8 +28,8 @@ class VideoPlayerItem: AVPlayerItem {
     self.createTracksLoadingTask()
   }
 
-  init?(videoSource: VideoSource) async throws {
-    guard let url = videoSource.uri else {
+  init?(videoSource: VideoSource, urlOverride: URL? = nil) async throws {
+    guard let url = urlOverride ?? videoSource.uri else {
       return nil
     }
     self.videoSource = videoSource